// packages/common/dsl/src/leader-thread/types.ts
export type DevtoolsContext =
  | {
      enabled: true;
      // syncBackendPullLatch: Effect.Latch
      // syncBackendPushLatch: Effect.Latch
      syncBackendLatch: Effect.Latch;
      syncBackendLatchState: SubscriptionRef.SubscriptionRef<{ latchClosed: boolean }>;
    }
  | {
      enabled: false;
    };